+++ tests/Giraffe.Tests/ModelBindingTests.fs
@@ -1274,6 +1274,46 @@ let ``BindModelAsync with JSON content returns correct result`` () =
     }
 
 [<Fact>]
+let ``BindModelAsync with QUERY JSON content returns correct result`` () =
+    let ctx = Substitute.For<HttpContext>()
+    mockJson ctx
+
+    let outputCustomer (c: Customer) = text (c.ToString())
+    let app = route "/auto" >=> bindModel<Customer> None outputCustomer
+
+    let contentType = "application/json"
+
+    let queryContent =
+        "{ \"name\": \"John Doe\", \"isVip\": true, \"birthDate\": \"1990-04-20\", \"balance\": 150000.5, \"loyaltyPoints\": 137 }"
+
+    let stream = new MemoryStream()
+    let writer = new StreamWriter(stream, Encoding.UTF8)
+    writer.Write queryContent
+    writer.Flush()
+    stream.Position <- 0L
+
+    let headers = HeaderDictionary()
+    headers.Add("Content-Type", StringValues(contentType))
+    headers.Add("Content-Length", StringValues(stream.Length.ToString()))
+    ctx.Request.ContentType.ReturnsForAnyArgs contentType |> ignore
+    ctx.Request.Method.ReturnsForAnyArgs "QUERY" |> ignore
+    ctx.Request.Path.ReturnsForAnyArgs(PathString("/auto")) |> ignore
+    ctx.Request.Headers.ReturnsForAnyArgs(headers) |> ignore
+    ctx.Response.Body <- new MemoryStream()
+    ctx.Request.Body <- stream
+
+    let expected =
+        "Name: John Doe, IsVip: true, BirthDate: 1990-04-20, Balance: 150000.50, LoyaltyPoints: 137"
+
+    task {
+        let! result = app (Some >> Task.FromResult) ctx
+
+        match result with
+        | None -> assertFailf "Result was expected to be %s" expected
+        | Some ctx -> Assert.Equal(expected, getBody ctx)
+    }
+
+[<Fact>]
 let ``BindModelAsync with XML content returns correct result`` () =
     let ctx = Substitute.For<HttpContext>()
     mockXml ctx